PM-3323: Split out the test connection manager.
